pyproject.toml
[project]
name = "exam-train"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "gradio>=5.42.0",
    "python-dotenv>=1.1.1",
]
